+++ catalogo-strumenti-backend/catalogo-backend-service/src/main/java/it/istat/mec/catalog/dao/BusinessServiceDao.java
@@ -7,13 +7,13 @@ import it.istat.mec.catalog.domain.BusinessService;
 
 
 @Repository
-public interface BusinessServiceDao extends JpaRepository<BusinessService, Long> {
+public interface BusinessServiceDao extends JpaRepository<BusinessService, Integer> {
 
 	@Override
 	List<BusinessService> findAll();
 	
 	@Override
-	Optional<BusinessService> findById(Long id);
+	Optional<BusinessService> findById(Integer id);
 
 	public void save(Optional<BusinessService> businessService);
