+++ catalogo-strumenti-backend/catalogo-backend-service/src/main/java/it/istat/mec/catalog/dao/DesktopApplicationDao.java
@@ -8,14 +8,14 @@ import it.istat.mec.catalog.domain.CatalogTool;
 import it.istat.mec.catalog.domain.DesktopApplication;
 
 @Repository
-public interface DesktopApplicationDao extends JpaRepository<DesktopApplication, Long> {
+public interface DesktopApplicationDao extends JpaRepository<DesktopApplication, Integer> {
 
 	@Override
 	List<DesktopApplication> findAll();
 	
 	
 	@Override
-	Optional<DesktopApplication> findById(Long id);
+	Optional<DesktopApplication> findById(Integer id);
 
 	public void save(Optional<DesktopApplication> tools);
